Integration with GameKeyListener - Renamed GameKeyListener to StackGameKeyListener. Created a deprecated version named GameKeyListener for now. - Renamed DefaultKeyMapper to BasicKeyMapper. - Created an Interface for GameKeyListeners. Poorly named until major version increment.
