+++ src/main/java/sugaEngine/input/GameKeyListenerInterface.java
@@ -1,0 +1,12 @@
+package sugaEngine.input;
+
+import javax.swing.*;
+import java.awt.event.KeyListener;
+
+/**
+ * The GameKeyListenerInterface is an object that responds whenever a key is pressed
+ */
+// todo rename to 'GameKeyListener'
+public interface GameKeyListenerInterface extends KeyListener {
+
+}
